chore: try to prevent circular import in the future
