+++ src/entity/notifications/NotificationPreferenceComment.ts
@@ -1,13 +1,13 @@
 import { ChildEntity, Column, ManyToOne } from 'typeorm';
 import { NotificationPreference } from './NotificationPreference';
 import { NotificationPreferenceType } from '../../notifications/common';
-import { Comment } from '../Comment';
+import type { Comment } from '../Comment';
 
 @ChildEntity(NotificationPreferenceType.Comment)
 export class NotificationPreferenceComment extends NotificationPreference {
   @Column({ type: 'text', default: null })
   commentId: string;
 
-  @ManyToOne(() => Comment, { lazy: true, onDelete: 'CASCADE' })
+  @ManyToOne('Comment', { lazy: true, onDelete: 'CASCADE' })
   comment: Promise<Comment>;
 }
