+++ src/entity/notifications/NotificationPreferencePost.ts
@@ -1,5 +1,5 @@
 import { ChildEntity, Column, ManyToOne } from 'typeorm';
-import { Post } from '../posts';
+import type { Post } from '../posts';
 import { NotificationPreferenceType } from '../../notifications/common';
 import { NotificationPreference } from './NotificationPreference';
 
@@ -8,6 +8,6 @@ export class NotificationPreferencePost extends NotificationPreference {
   @Column({ type: 'text', default: null })
   postId: string;
 
-  @ManyToOne(() => Post, { lazy: true, onDelete: 'CASCADE' })
+  @ManyToOne('Post', { lazy: true, onDelete: 'CASCADE' })
   post: Promise<Post>;
 }
