+++ src/entity/notifications/NotificationPreferenceSource.ts
@@ -1,5 +1,5 @@
 import { ChildEntity, Column, ManyToOne } from 'typeorm';
-import { Source } from '../Source';
+import type { Source } from '../Source';
 import { NotificationPreferenceType } from '../../notifications/common';
 import { NotificationPreference } from './NotificationPreference';
 
@@ -8,6 +8,6 @@ export class NotificationPreferenceSource extends NotificationPreference {
   @Column({ type: 'text', default: null })
   sourceId: string;
 
-  @ManyToOne(() => Source, { lazy: true, onDelete: 'CASCADE' })
+  @ManyToOne('Source', { lazy: true, onDelete: 'CASCADE' })
   source: Promise<Source>;
 }
